+++ com.unity.formats.alembic/Editor/Importer/AlembicBuildPostProcess.cs
@@ -41,6 +41,11 @@ namespace UnityEditor.Formats.Alembic.Importer
 
             foreach (var files in FilesToCopy)
             {
+                if (!File.Exists(files.Key))
+                {
+                    continue;
+                }
+
                 var dir = Path.GetDirectoryName(files.Value);
                 if (dir != null && !Directory.Exists(dir))
                 {
@@ -95,6 +100,10 @@ namespace UnityEditor.Formats.Alembic.Importer
         {
             streamPlayer.StreamDescriptor = streamPlayer.StreamDescriptor.Clone();// make a copy
             var srcPath = streamPlayer.StreamDescriptor.PathToAbc;
+            if (string.IsNullOrEmpty(srcPath))
+            {
+                return;
+            }
 
             // Avoid name collisions by hashing the full path
             var hashedFilename = HashSha1(srcPath) + ".abc";